added imgs for pics and backgrounds, made navbar and aligned right, add loremipsum txt to body of homepage, created folders in assets to organize images, corrected mapping to images, added font families.
