+++ application/src/main/resources/static/js/react-client/src/assets/background/dark-mode-background.png
