+++ application/src/main/resources/static/js/react-client/src/assets/background/light-mode-background.png
