+++ application/src/main/resources/static/js/react-client/src/assets/logo/JAC-Game-dark.png
