+++ application/src/main/resources/static/js/react-client/src/pages/contact-screen/index.js
@@ -1,14 +1,14 @@
-import JacLogo from '../../assets/JAC-Game-light.png';
+import JacLogoL from '../../assets/logo/JAC-Game-light.png';
 import React from 'react';
-import alanIcon from '../../assets/alan.png';
-import cliffIcon from '../../assets/cliff.png';
-import jordanIcon from '../../assets/jordan.png';
+import alanIcon from '../../assets/contact/alan.png';
+import cliffIcon from '../../assets/contact/cliff.png';
+import jordanIcon from '../../assets/contact/jordan.png';
 import style from './style.module.scss';
 
 const ContactScreen = () => {
     return (
         <div className={style.ContactScreen}>
-            <img src={JacLogo} alt='JAC Logo' />
+            <img src={JacLogoL} alt='JAC Logo' />
             <h2>Contact Us!</h2>
             <img src={cliffIcon} alt='Cliff Jenkins CEO' />
             <img src={alanIcon} alt='Alan Kostrick VP' />
