+++ application/src/main/resources/static/js/react-client/src/pages/home-screen/style.module.scss
@@ -1,5 +1,5 @@
 body {
-    background-image: url(../../assets/light-mode-background.png);
+    background-image: url(../../assets/background/light-mode-background.png);
     background-size: cover;
 
     & p {
