+++ application/src/main/resources/static/js/react-client/src/style.module.scss
@@ -1,3 +1,12 @@
+
+@import url('https://fonts.googleapis.com/css2?family=Press+Start+2P&family=Roboto&display=swap');
+
+
+body {
+
+    font-family: 'Roboto', sans-serif;
+}
+
 // .page {
 //     // background-color: #ddd;
 //     height: 1000px;
